# src/main/resources/openapi/openapi.yml
openapi: 3.0.0
info:
  title: A_iDa_012
  version: 1.3.43
paths:
  /account/{accountNumber}/transactions:
    get:
      parameters:
        - name: accountNumber
          in: path
          required: true
          description: "Номер счета"
          schema:
            type: string
        # - name: curFormat
        #   in: query
        #   required: false
        #   description: "Формат операции по валютному счету"
        #   schema:
        #     type: string
        - name: statementDate
          in: query
          description: "Дата запрашиваемой выписки"
          required: true
          schema:
            type: string
            format: date-time
      responses:
        '200':
          description: successful operation
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Transactions"
        '400':
          description: Invalid username supplied
        '404':
          description: Not Found
          
components:
  schemas:
    Transactions:
        properties:
          transactions:
            type: array
            items: 
              $ref: '#/components/schemas/Transaction'
        required:
          - transactions

    Transaction:  
      type: object
      properties:
        amount:
          $ref: '#/components/schemas/Amount'
        correspondingAccount:
          type: string
          pattern: '\d{20}$'
        swiftTransfer:
          $ref: '#/components/schemas/SwiftTransfer'
        uuid:
          type: string
          format: uuid
        transactionId:
          type: integer
          format: int64
          pattern: '^\d{13}$'
      required:
        - correspondingAccount
        - uuid
        - transactionId
            
          
    Amount:
      type: object
      properties:
        amount:
          type: number
          format: float
          minimum: 0.01
          pattern: '^\d*.\d{1,2}$'
        currencyName:
          type: string
          pattern: '^\w[A-Z]{3}$'
      required:
        - amount
        - currencyName
          
          
    SwiftTransfer:
      type: object
      properties:
        bankOperationCode:
          type: string
          pattern: '^\w{4}$'
        beneficiaryBankAccount:
          type: string
          format: password 
          maxLength: 11
        beneficiaryBankName:
          type: string
        beneficiaryBankOption:  
          type: string
          pattern: '^\w[A-Z]{1}$'
        beneficiaryCustomerAccount:
          type: string
          format: password
        beneficiaryCustomerName:
          type: string
        detailsOfCharges:
          type: string
          pattern: '^\w[A-Z]{3}$'
        exchangeRate:
          type: string
          format: float
        instructedAmount:
          type: string
        instructionCode:
          type: string
          pattern: '^\w[A-Z]{4}$'
        intermediaryBankAccount:
          type: string
          format: password
          pattern: '^\w[A-Z]{8}$'
        intermediaryBankName:
          type: string
        intermediaryBankOption:
          type: string
          pattern: '^\w[A-Z]{1}$'
        messageDestinator:
          type: string
        messageIdentifier:
          type: string
        essageOriginator:
          type: string
        messageReceiveTime:
          type: string
          format: date-time
        messageSendTime:
          type: string
          format: date-time
        messageType:
          type: string
        orderingCustomerAccount:
          type: string
          format: password
        orderingCustomerName:
          type: string
        orderingCustomerOption:
          type: string
          pattern: '^\w[A-Z]{1}$'
        orderingInstitutionAccount:
          type: string
        orderingInstitutionName:
          type: string
        orderingInstitutionOption:
          type: string
          pattern: '^\w[A-Z]{1}$'
        receiverCharges:
          type: string
          pattern: '^\w[A-Z]{3}\d+,\d{1,2}$'
        receiverCorrespondentAccount:
          type: string
          format: password
        receiverCorrespondentName:
          type: string
        receiverCorrespondentOption:
          type: string
          pattern: '^\w[A-Z]{1}$'
        regulatoryReporting:
          type: string
        remittanceInformation:
          type: string
        senderCharges:
          type: string
          pattern: '^\w[A-Z]{3}\d+,\d{1,2}$'
        senderCorrespondentAccount:
          type: string
          format: password
        senderCorrespondentName:
          type: string
        senderCorrespondentOption:
          type: string
          pattern: '^\w[A-Z]{1}$'
        senderToReceiverInformation:
          type: string
        transactionReferenceNumber:
          type: string
        transactionRelatedReference:
          type: string
        transactionTypeCode:
          type: string
        urgent:
          type: string
          pattern: '^\w[A-Z]{6}$'
        valueDateCurrencyInterbankSettledAmount:
          type: string
      required:
        - bankOperationCode
          
      
                        
                        
                        
                    
        